Support booting on ext4 for zuma devices
When a user opts into 16K developer option, we would need to convert the
/data and /metadata partition into ext4. Add necessary fstab entries for
ext4 so that zuma devices can boot on ext4.

This CL does not automatically switch existing devices to ext4, the
newly added fstab entries are intentionally marked as "non-formattable",
so that we don't accidentally format a wiped device as ext4. This CL
merely allows the device to boot if the /data partition is already
formatted as ext4.

Test: adb shell cmd recovery wipe ext4 , make sure device boots
Bug: 293313353
Change-Id: I3a2a2e9d09cdea884f58b509a06c6829938dc369