usb: handle displayport debounce timer empty reads
If the device is unplugged then plugged before
mDisplayPortDebounceTimer triggers, then the new
displayPortPollWork could process the trigger
in the intial epoll handling loop. Because the
timer can be rearmed, the timerfd read results
in -EAGAIN because no data is ready, but stale
data was ready to begin with.

Disarm the debounce timer when displayPortPollWork
closes to prevent new work threads from reading
timer when not ready. Also log debounce read
errors as a warning and not an error.

Test: manual test on device
Bug: 328553922
(cherry-picked from 94c9c1ac86f398c22ddc446f17bd7ab3fddb675c)
Change-Id: Icf3dbe929eb9f0520941a4efe3dcea79a71d46c3
Signed-off-by: RD Babiera <rdbabiera@google.com>